hero enhancements
